\name{PeakSeg-package}
\alias{PeakSeg-package}
\alias{PeakSeg}
\docType{package}
\title{
\packageTitle{PeakSeg}
}
\description{
\packageDescription{PeakSeg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PeakSeg}
\packageIndices{PeakSeg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{PeakSeg}

Maintainer: \packageMaintainer{PeakSeg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
	vec=rnorm(10000)
	seg.data=PeakSeg(vec,winnum=3,step=20)
}
